refactor: service 패키지 분리 및 이름 수정
