+++ src/main/java/chess/Application.java
@@ -1,10 +1,10 @@
 package chess;
 
 import chess.controller.ChessGameController;
-import chess.dao.ChessDBService;
 import chess.dao.ChessGameDao;
 import chess.dao.ConnectionGenerator;
 import chess.dao.GameInformationDao;
+import chess.service.ChessService;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -13,7 +13,7 @@ public class Application {
 
     public static void main(String[] args) {
         ChessGameController chessGameController = new ChessGameController(new InputView(), new OutputView(),
-                new ChessDBService(
+                new ChessService(
                         new GameInformationDao(ConnectionGenerator.from(CONFIGURATION_FILE_NAME)),
                         new ChessGameDao(ConnectionGenerator.from(CONFIGURATION_FILE_NAME))
                 ));
